Use GSON instead of Jackson to avoid affecting downstream projects
Having Jackson annotations on the types in the tooling-spi project
caused downstream projects implementing the SPI to have a couple of
Jackson JARs and their dependencies on the classpath, including a
different XML API implementation. Switching to GSON avoids that and also
lets us get rid of the extra dependency on json-sanitizer in
tooling-core because GSON already escapes problematic HTML characters.