Allow Git URLs without a specified schema (#192)
Prior to this commit, the XML schema was too restrictive and only 
allowed actual URIs in the `originUrl` attribute of the `git:repository`
element. Since Git allows values like `git@github.com:org/repo` (without
the `ssh://` prefix to make it an URI), we schema now allows that 
as well.