Add support for sending events to a `Writer` directly (#728)
Prior to this change, the `Event` class only accepted a file, but it 
would be much more generic to use a writer and the documentation even 
mention that use
case.

This now enhances the API with a new way to create a (event) writer that
prints to the (Java) writer directly.

---------

Co-authored-by: Marc Philipp <mail@marcphilipp.de>